% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmfit.R
\name{dlmfit}
\alias{dlmfit}
\title{Fits a bayesian dynamic linear model (DLM) of the (univariate) series y to the
 covariates covars}
\usage{
dlmfit(y, covars, W.disc.factor = 0.9, v = 1, prior.dist = "N",
  prior.mean = 0, prior.cov = 1e+08, ng.alpha = 1, ng.beta = 1,
  calc.hist.vars = TRUE)
}
\arguments{
\item{y}{is a vector of equally-spaced observations of the dependent series y}

\item{covars}{is a matrix of equally-spaced observations of covariates. Each covariate
is contained in a column. The number of rows of the covars matrix must be equal
to the length of the vector y}

\item{W.disc.factor}{is the discount factor applied in the estimation of W}

\item{V}{is blah}
}
\description{
\code{dlmtfit} fits a dynamic linear model of the time series y to the
covariate time series in covars.
}
