% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmfit.R
\name{is.scalar}
\alias{is.scalar}
\title{Checks whether a provided value is a scalar.}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{is an object of unknown type}
}
\description{
\code{is.scalar} returns true if the provided argument is a scalar and false
otherwise.
}
\examples{
is.scalar('A')
 is.scalar(matrix(5))
 is.scalar(sum)
}
